v0.7c 60 easy tests are fully working... (also rewrote some printing-stuff)
introduced boolean "printDebug" in solve()-method-signature in Sudoku. So I can tell the solver, if I wan't to have extensive printing or not.

This version of sudoku can already solve simple sudokus, of various sizes. But it doesn't make assumptions, if it cannot solve it with the already implemented algorithms. That's one of the next goals.